Refactored files to seperate fucntional code from styles and util.
