+++ src/stories/Filter/FilterPage/index.ts
@@ -1,0 +1,3 @@
+import FilterPage from './FilterPage';
+
+export default FilterPage;
