+++ src/stories/Filter/components/ContextSwitchMenu/index.ts
@@ -1,0 +1,3 @@
+import ContextSwitchMenu from './ContextSwitchMenu';
+
+export default ContextSwitchMenu;
