+++ src/stories/Filter/components/FilterBar/index.ts
@@ -1,0 +1,3 @@
+import FilterBar from './FilterBar';
+
+export default FilterBar;
