+++ src/stories/Filter/components/FilterJSONConfirmationModal/FilterJSONConfirmationModal.tsx
@@ -1,43 +1,7 @@
 import { FC } from 'react';
-import styled from '@emotion/styled';
 import { FilterJSONConfirmationModalProps } from 'src/Filter/types';
 import Button from 'src/Button/Button';
-
-const ModalContainer = styled.div<FilterJSONConfirmationModalProps>(({ modalVisible = false }) => ({
-    display: modalVisible ? 'block' : 'none',
-    position: 'fixed',
-    zIndex: 1,
-    paddingTop: '100px',
-    left: '0',
-    top: '0',
-    width: '100%',
-    height: '100%',
-    overflow: 'auto',
-    backgroundColor: ['rgb(0,0,0)', 'rgba(0,0,0,0.4)']
-}));
-
-const ModalContent = styled.div({
-    backgroundColor: '#fefefe',
-    margin: '10% auto auto',
-    padding: '20px',
-    border: '1px solid #888',
-    width: '80%'
-});
-
-const ButtonWrapper = styled.div({
-    display: 'flex',
-    justifyContent: 'space-between',
-    marginTop: '1em',
-    button: {
-        flex: 1
-    },
-    'button:first-of-type': {
-        marginRight: 2
-    },
-    'button:last-of-type': {
-        marginLeft: 2
-    }
-});
+import { ButtonWrapper, ModalContainer, ModalContent } from './filterJSONConfirmationModalStyles';
 
 const FilterJSONConfirmationModal: FC<FilterJSONConfirmationModalProps> = ({
     modalVisible,
