+++ src/stories/Filter/components/FilterJSONConfirmationModal/index.ts
@@ -1,0 +1,3 @@
+import FilterJSONConfirmationModal from './FilterJSONConfirmationModal';
+
+export default FilterJSONConfirmationModal;
