+++ src/stories/Filter/components/FilterJSONInput/FilterJSONInput.tsx
@@ -1,28 +1,7 @@
 import { FC, useEffect, useState } from 'react';
-import styled from '@emotion/styled';
-import TextArea from 'src/TextArea/TextArea';
 import Button from 'src/Button/Button';
 import { FilterJSONInputProps } from 'src/Filter/types';
-
-const JSONEditor = styled(TextArea)({
-    marginTop: '1em',
-    width: 231
-});
-
-const ButtonWrapper = styled.div({
-    display: 'flex',
-    justifyContent: 'space-between',
-    marginTop: '1em',
-    button: {
-        flex: 1
-    },
-    'button:first-of-type': {
-        marginRight: 2
-    },
-    'button:last-of-type': {
-        marginLeft: 2
-    }
-});
+import { ButtonWrapper, JSONEditor } from './filterJSONInputStyles';
 
 const JSON_INDENT_SPACES = 4;
