+++ src/stories/Filter/components/FilterJSONInput/index.ts
@@ -1,0 +1,3 @@
+import FilterJSONInput from './FilterJSONInput';
+
+export default FilterJSONInput;
